Merged PR 125: Canon: draft high-level user-facing std lib operations
This PR drafts higher level operations for use by users in writing algorithms in Q♭. Some of these we might want to later suggest for inclusion in the Q♭ standard library itself, but this PR separates the high-level operations acting on operations such that we can develop on them independently.

Cherry picked from Pull request 122.