Merged PR 221: Miscellaneous new operations
This PR adds an assortment of new operations that I found useful in developing the quantum process tomography demo. In particular, the new operations ``MReset*`` for ``*`` ∈ {`X`, `Y`, `Z`} ensure the post-measurement condition that the measured qubit is in the |0〉 state, appropriate for deallocation at the end of a `using` block. Similarly, the new `Identity` operation does nothing at all, such that it can be the last statement in a `using` block, which is necessary if you want to ignore the result of an operation invoked earlier using `let`.