Merged PR 244: Initial draft of style guide.
Since we're getting to have a variety of different coding styles, I thought it would be good moving forward to agree on a few conventions. I've added an initial draft to this effect, based on what seems to be emergent from our code so far. Please treat all of these entries as suggestions only; we should decide upon a set of conventions that works as well as possible for everyone, after all!