Merged PR 488: Fixes for a variety of different docgen issues.
This PR uses !487 to generate YAML summaries for almost all remaining API documentation that is currently in Libraries. The only remaining failures reported by docgen.exe are in functions and operations whose return types are singleton tuples of generic types denoted without parens. These callables do not yet parse due to a known issue. That's also not to say that the generated YAML is correct, only that no errors are raised by the generator.

(Many of these fixes are removing trailing whitespace in /// comments, as that seems to sometimes trip up the /// parser.)